requirements.txt
fastapi
uvicorn
pydantic
faiss-cpu
strawberry-graphql
ollama
python-multipart
pyyaml
requests
pyyaml
sentence-transformers
slowapi==0.1.5
protobuf

